feat: all serialisation to/from hex strings should now give bigendian strings
