+++ lib/src/lib.rs
@@ -6,6 +6,7 @@ pub mod serialization;
 pub mod utils;
 pub mod validate_certificate;
 pub mod wallets;
+pub mod hex_conversion;
 
 pub mod initialisation {
     use crate::blockchain_client::BlockchainClientConnection;
