+++ lib/src/serialization.rs
@@ -1,10 +1,10 @@
+use crate::hex_conversion::HexConvertible;
 use crate::merkle_trees::trees::MerkleTreeError;
 use ark_bn254::Fr;
 use ark_serialize::{
     CanonicalDeserialize, CanonicalSerialize, Compress, SerializationError, Validate,
 };
 use hex;
-use hex::encode;
 use mongodb::error::Error as MongoError;
 use serde::{ser::Serialize, Deserialize, Deserializer, Serializer};
 use std::fmt::Debug;
@@ -30,6 +30,7 @@ where
     Ok(a)
 }
 
+/// Serializes an arkworks element to a hex string using a bigendian representation.
 pub fn ark_se_hex<S, A: CanonicalSerialize>(a: &A, s: S) -> Result<S::Ok, S::Error>
 where
     S: serde::Serializer,
@@ -37,7 +38,8 @@ where
     let mut bytes = vec![];
     a.serialize_with_mode(&mut bytes, Compress::Yes)
         .map_err(serde::ser::Error::custom)?;
-    let hex_str = encode(&bytes);
+    bytes.reverse(); // Convert to big-endian
+    let hex_str = bytes.to_hex_string();
     s.serialize_str(&hex_str)
 }
 
@@ -46,8 +48,8 @@ where
     D: serde::de::Deserializer<'de>,
 {
     let hex_str: &str = serde::de::Deserialize::deserialize(data)?;
-    let bytes = hex::decode(hex_str).map_err(serde::de::Error::custom)?;
-
+    let mut bytes = Vec::<u8>::from_hex_string(hex_str).map_err(serde::de::Error::custom)?;
+    bytes.reverse(); // Convert to little-endian (which is the expected format for arkworks)
     let a = A::deserialize_with_mode(bytes.as_slice(), Compress::Yes, Validate::Yes)
         .map_err(serde::de::Error::custom)?;
     Ok(a)
@@ -190,8 +192,10 @@ impl<'de> Deserialize<'de> for FrWrapperPadded {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::hex_conversion::HexConvertible;
 
+    use super::*;
+    use ark_bn254::Fr as Fr254;
     use std::str::FromStr;
     #[test]
     fn test_ark_se_and_ark_de_1() {
@@ -217,9 +221,12 @@ mod tests {
     }
     #[test]
     fn test_ark_se_hex_and_ark_de_hex() {
-        let element = Fr::from_str("10").unwrap();
+        // 43981 is the decimal representation of 0xABCD
+        let element = Fr254::from_str("43981").unwrap();
         let wrapper = FrWrapperhex(element);
         let serialized = serde_json::to_string(&wrapper).unwrap();
+        dbg!(&serialized);
+        dbg!(element.to_hex_string());
         let deserialized: FrWrapperhex = serde_json::from_str(&serialized).unwrap();
         let deserialized_element = deserialized.0;
         assert_eq!(element, deserialized_element);
