+++ nightfall_client/src/drivers/rest/commitment.rs
@@ -1,7 +1,7 @@
 use log::trace;
 use warp::{hyper::StatusCode, path, reject, reply, Filter, Reply};
 
-use crate::domain::entities::HexConvertible;
+use lib::hex_conversion::HexConvertible;
 use crate::driven::db::mongo::CommitmentEntry;
 use crate::initialisation::get_db_connection;
 use crate::ports::db::CommitmentDB;
