+++ nightfall_client/src/drivers/rest/utils.rs
@@ -1,6 +1,7 @@
 //! File contains utility functions used by the REST API, such as ones for converting from erc address and token id to
 //! Nightfall token id.
-use crate::domain::{entities::HexConvertible, error::ConversionError};
+use crate::domain::error::ConversionError;
+use lib::hex_conversion::HexConvertible;
 use ark_bn254::Fr as Fr254;
 use ark_ff::{BigInteger, PrimeField};
 use ethers::{
